requirements.txt
# LINE Bot SDK
line-bot-sdk==3.8.0

# OpenAI SDK
openai>=1.12.0

# HTTP Client
requests>=2.31.0

# Web Framework (for Vercel)
Flask>=3.0.0

# Environment Variables
python-dotenv>=1.0.0

# Vercel KV (Redis)
redis>=5.0.0

# Image Processing
Pillow>=10.0.0

# Testing
pytest>=7.4.0
pytest-mock>=3.12.0
